Adding Libc Call Logging Capabilities
Monguard is now able to count the number of calls have been made to the Monguard trampoline
This is especially useful to count libc calls made in a particular subgraph of a calltree.
Simply use the new library functions which will be preloaded, start_libcall_count and
end_libcall_count to flank the function call you wish to track libc calls in.
Additionally, to enable also change inc/config.h and define _LIBC_COUNT.